Reworked wallet inheritance hierarchy
